# routes/employees.js
const express = require('express');
const router = express.Router();
const employeesController = require('../controllers/employees');

/**
 * @swagger
 * tags:
 *   name: Employees
 *   description: API for employees management
 */

/**
 * @swagger
 * /employees:
 *   get:
 *     summary: Get all employees
 *     tags: [Employees]
 *     responses:
 *       200:
 *         description: List of employees
 *         content:
 *           application/json:
 *             schema:
 *               type: array
 *               items:
 *                 type: object
 *                 properties:
 *                   _id:
 *                     type: string
 *                   firstName:
 *                     type: string
 *                   lastName:
 *                     type: string
 *                   position:
 *                     type: string
 *                   department:
 *                     type: string
 */
router.get('/', employeesController.getAll);

/**
 * @swagger
 * /employees/{id}:
 *   get:
 *     summary: Get an employee by ID
 *     tags: [Employees]
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema:
 *           type: string
 *         description: Employee ID
 *     responses:
 *       200:
 *         description: Employee found
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 _id:
 *                   type: string
 *                 fullName:
 *                   type: string
 *                 phoneNumber:
 *                   type: string
 *                 hireDate:
 *                   type: string
 *                 department:
 *                   type: string
 *                 employmentStatus:
 *                   type: string
 *                 role:
 *                   type: string
 *                 address:
 *                   type: string
 *       404:
 *         description: Employee not found
 */
router.get('/:id', employeesController.getSingle);

/**
 * @swagger
 * /employees:
 *   post:
 *     summary: Create a new employee
 *     tags: [Employees]
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             required:
 *               - fullName
 *               - phoneNumber
 *               - hireDate
 *               - department
 *               - employmentStatus
 *               - role
 *               - address
 *             properties:
 *               fullName:
 *                 type: string
 *               phoneNumber:
 *                 type: string
 *               hireDate:
 *                 type: string
 *               department:
 *                 type: string
 *               employmentStatus:
 *                 type: string
 *               role:
 *                 type: string
 *               address:
 *                 type: string
 *     responses:
 *       201:
 *         description: Employee created
 *       400:
 *         description: Invalid input
 */
router.post('/', employeesController.createEmployee);

/**
 * @swagger
 * /employees/{id}:
 *   put:
 *     summary: Update an employee
 *     tags: [Employees]
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema:
 *           type: string
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             properties:
 *               fullName:
 *                 type: string
 *               phoneNumber:
 *                 type: string
 *               hireDate:
 *                 type: string
 *               department:
 *                 type: string
 *               employmentStatus:
 *                 type: string
 *               role:
 *                 type: string
 *               address:
 *                 type: string
 *     responses:
 *       200:
 *         description: Employee updated
 *       404:
 *         description: Employee not found
 */
router.put('/:id', employeesController.updateEmployee);

/**
 * @swagger
 * /employees/{id}:
 *   delete:
 *     summary: Delete an employee
 *     tags: [Employees]
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema:
 *           type: string
 *     responses:
 *       200:
 *         description: Employee deleted
 *       404:
 *         description: Employee not found
 */
router.delete('/:id', employeesController.deleteEmployee);

module.exports = router;